remove the gw_labkey tables only if they exist in the database
SVN r43127 |2016-03-29 21:43:36 +0000

+++ hdrl/src/org/labkey/hdrl/HDRLContainerListener.java
@@ -19,12 +19,16 @@ package org.labkey.hdrl;
 import org.jetbrains.annotations.NotNull;
 import org.labkey.api.data.Container;
 import org.labkey.api.data.ContainerManager.ContainerListener;
+import org.labkey.api.data.DatabaseTableType;
 import org.labkey.api.data.DbSchema;
 import org.labkey.api.data.DbScope;
 import org.labkey.api.data.SQLFragment;
+import org.labkey.api.data.SchemaTableInfo;
 import org.labkey.api.data.SqlExecutor;
 import org.labkey.api.security.User;
 import org.labkey.api.util.ContainerUtil;
+import org.labkey.hdrl.query.LabWareQuerySchema;
+import org.labkey.remoteapi.assay.Data;
 
 import java.util.Collections;
 import java.util.Collection;
@@ -42,27 +46,29 @@ public class HDRLContainerListener implements ContainerListener
     public void containerDeleted(Container c, User user)
     {
         DbSchema hdrl = HDRLSchema.getInstance().getSchema();
+        DbSchema labware = new LabWareQuerySchema(user, c).getSchema();
         try (DbScope.Transaction transaction = hdrl.getScope().ensureTransaction())
         {
-            SqlExecutor executor = new SqlExecutor(hdrl.getScope());
-            SQLFragment deleteSQL = getLabWareDeleteSQL("X_LK_OUTBD_SPECIMENS", c);
-            executor.execute(deleteSQL);
-            deleteSQL = getLabWareDeleteSQL("X_LK_OUTBD_REQUESTS", c);
-            executor.execute(deleteSQL);
-            deleteSQL = getLabWareDeleteSQL("X_LK_INBND_SPECIMENS", c);
-            executor.execute(deleteSQL);
-            deleteSQL = getLabWareDeleteSQL("X_LK_INBND_REQUESTS", c);
-            executor.execute(deleteSQL);
+            deleteLabWareTable(labware, "X_LK_OUTBD_SPECIMENS", c);
+            deleteLabWareTable(labware, "X_LK_OUTBD_REQUESTS", c);
+            deleteLabWareTable(labware, "X_LK_INBND_SPECIMENS", c);
+            deleteLabWareTable(labware, "X_LK_INBND_REQUESTS", c);
+
             ContainerUtil.purgeTable(hdrl.getTable("InboundRequest"), c, null);
             transaction.commit();
         }
     }
 
-    private SQLFragment getLabWareDeleteSQL(String labWareTable, Container container)
+    private void deleteLabWareTable(DbSchema labware, String labWareTable, Container container)
     {
-        SQLFragment deleteSQL = new SQLFragment("DELETE FROM gw_labkey.").append(labWareTable).append(" WHERE batch_id in (SELECT RequestId from hdrl.InboundRequest WHERE Container = ?)");
-        deleteSQL.add(container.getId());
-        return deleteSQL;
+        SqlExecutor executor = new SqlExecutor(labware.getScope());
+        SchemaTableInfo tableInfo = labware.getTable("X_LK_OUTBD_SPECIMENS");
+        if (tableInfo != null && tableInfo.getTableType() != DatabaseTableType.NOT_IN_DB)
+        {
+            SQLFragment deleteSQL = new SQLFragment("DELETE FROM gw_labkey.").append(labWareTable).append(" WHERE batch_id in (SELECT RequestId from hdrl.InboundRequest WHERE Container = ?)");
+            deleteSQL.add(container.getId());
+            executor.execute(deleteSQL);
+        }
     }
 
     @Override
